+++ packages/core-nodes-gmail/src/nodes/ReplyToGmailMessageNode.ts
@@ -1,12 +1,16 @@
 import type { Item, RunnableNode, RunnableNodeExecuteArgs } from "@codemation/core";
 import { inject, node } from "@codemation/core";
 import { GmailReplyToMessageService } from "../services/GmailReplyToMessageService";
-import type { ReplyToGmailMessage, ReplyToGmailMessageOutputJson } from "./ReplyToGmailMessage";
+import type {
+  ReplyToGmailMessage,
+  ReplyToGmailMessageInputJson,
+  ReplyToGmailMessageOutputJson,
+} from "./ReplyToGmailMessage";
 
 @node({ packageName: "@codemation/core-nodes-gmail" })
 export class ReplyToGmailMessageNode implements RunnableNode<
   ReplyToGmailMessage,
-  unknown,
+  ReplyToGmailMessageInputJson,
   ReplyToGmailMessageOutputJson
 > {
   readonly kind = "node" as const;
@@ -16,9 +20,13 @@ export class ReplyToGmailMessageNode implements RunnableNode<
     @inject(GmailReplyToMessageService) private readonly gmailReplyToMessageService: GmailReplyToMessageService,
   ) {}
 
-  async execute(args: RunnableNodeExecuteArgs<ReplyToGmailMessage, unknown>): Promise<Item> {
+  async execute(args: RunnableNodeExecuteArgs<ReplyToGmailMessage, ReplyToGmailMessageInputJson>): Promise<Item> {
     return {
-      json: await this.gmailReplyToMessageService.reply(args.ctx),
+      json: await this.gmailReplyToMessageService.reply({
+        input: args.input,
+        item: args.item,
+        ctx: args.ctx,
+      }),
     };
   }
 }